Rename algorithm to alg
