+++ js/src/array/macro.rs
@@ -10,74 +10,74 @@ macro_rules! impl_geometry_array {
                 &self,
                 transform: BroadcastableAffine,
             ) -> WasmResult<GeometryArray> {
-                use geoarrow::algorithm::geo::affine_transform;
+                use geoarrow::alg::geo::affine_transform;
                 Ok(GeometryArray(affine_transform(&self.into(), transform.0)?))
             }
 
             #[wasm_bindgen]
             pub fn area(&self) -> WasmResult<FloatArray> {
-                use geoarrow::algorithm::geo::area;
+                use geoarrow::alg::geo::area;
                 Ok(FloatArray(area(self.into())?))
             }
 
             #[wasm_bindgen]
             pub fn center(&self) -> WasmResult<PointArray> {
-                use geoarrow::algorithm::geo::center;
+                use geoarrow::alg::geo::center;
                 Ok(PointArray(center(&self.into())?))
             }
 
             #[wasm_bindgen]
             pub fn centroid(&self) -> WasmResult<PointArray> {
-                use geoarrow::algorithm::geo::centroid;
+                use geoarrow::alg::geo::centroid;
                 Ok(PointArray(centroid(&self.into())?))
             }
 
             #[wasm_bindgen]
             pub fn convex_hull(&self) -> WasmResult<PolygonArray> {
-                use geoarrow::algorithm::geo::convex_hull;
+                use geoarrow::alg::geo::convex_hull;
                 Ok(PolygonArray(convex_hull(&self.into())?))
             }
 
             #[wasm_bindgen]
             pub fn euclidean_length(&self) -> WasmResult<FloatArray> {
-                use geoarrow::algorithm::geo::euclidean_length;
+                use geoarrow::alg::geo::euclidean_length;
                 Ok(FloatArray(euclidean_length(&self.into())?))
             }
 
             #[wasm_bindgen]
             pub fn geodesic_area(&self) -> WasmResult<FloatArray> {
-                use geoarrow::algorithm::geo::geodesic_area_unsigned;
+                use geoarrow::alg::geo::geodesic_area_unsigned;
                 Ok(FloatArray(geodesic_area_unsigned(&self.into())?))
             }
 
             #[wasm_bindgen]
             pub fn geodesic_area_signed(&self) -> WasmResult<FloatArray> {
-                use geoarrow::algorithm::geo::geodesic_area_signed;
+                use geoarrow::alg::geo::geodesic_area_signed;
                 Ok(FloatArray(geodesic_area_signed(&self.into())?))
             }
 
             #[wasm_bindgen]
             pub fn geodesic_length(&self) -> WasmResult<FloatArray> {
-                use geoarrow::algorithm::geo::geodesic_length;
+                use geoarrow::alg::geo::geodesic_length;
                 Ok(FloatArray(geodesic_length(&self.into())?))
             }
 
             #[wasm_bindgen]
             pub fn haversine_length(&self) -> WasmResult<FloatArray> {
-                use geoarrow::algorithm::geo::haversine_length;
+                use geoarrow::alg::geo::haversine_length;
                 Ok(FloatArray(haversine_length(&self.into())?))
             }
 
             #[wasm_bindgen]
             pub fn is_empty(&self) -> WasmResult<BooleanArray> {
-                use geoarrow::algorithm::geo::is_empty;
+                use geoarrow::alg::geo::is_empty;
                 Ok(BooleanArray(is_empty(&self.into())?))
             }
 
             #[cfg(feature = "geodesy")]
             #[wasm_bindgen]
             pub fn reproject_rs(&self, definition: &str) -> WasmResult<GeometryArray> {
-                use geoarrow::algorithm::geodesy::{reproject, Direction};
+                use geoarrow::alg::geodesy::{reproject, Direction};
                 Ok(GeometryArray(reproject(
                     &self.into(),
                     definition,
@@ -91,7 +91,7 @@ macro_rules! impl_geometry_array {
                 angle: BroadcastableFloat,
                 origin: TransformOrigin,
             ) -> WasmResult<GeometryArray> {
-                use geoarrow::algorithm::geo::rotate;
+                use geoarrow::alg::geo::rotate;
                 Ok(GeometryArray(rotate(&self.into(), angle.0, origin.0)?))
             }
 
@@ -102,7 +102,7 @@ macro_rules! impl_geometry_array {
                 yfact: BroadcastableFloat,
                 origin: TransformOrigin,
             ) -> WasmResult<GeometryArray> {
-                use geoarrow::algorithm::geo::scale;
+                use geoarrow::alg::geo::scale;
                 Ok(GeometryArray(scale(
                     &self.into(),
                     xfact.0,
@@ -113,7 +113,7 @@ macro_rules! impl_geometry_array {
 
             #[wasm_bindgen]
             pub fn signed_area(&self) -> WasmResult<FloatArray> {
-                use geoarrow::algorithm::geo::signed_area;
+                use geoarrow::alg::geo::signed_area;
                 Ok(FloatArray(signed_area(self.into())?))
             }
 
@@ -124,7 +124,7 @@ macro_rules! impl_geometry_array {
                 y_degrees: BroadcastableFloat,
                 origin: TransformOrigin,
             ) -> WasmResult<GeometryArray> {
-                use geoarrow::algorithm::geo::skew;
+                use geoarrow::alg::geo::skew;
                 Ok(GeometryArray(skew(
                     &self.into(),
                     x_degrees.0,
@@ -146,7 +146,7 @@ macro_rules! impl_geometry_array {
                 x_offset: BroadcastableFloat,
                 y_offset: BroadcastableFloat,
             ) -> WasmResult<GeometryArray> {
-                use geoarrow::algorithm::geo::translate;
+                use geoarrow::alg::geo::translate;
                 log!("{:?}", &self.0);
                 Ok(GeometryArray(translate(
                     &self.into(),
@@ -157,7 +157,7 @@ macro_rules! impl_geometry_array {
 
             #[wasm_bindgen]
             pub fn vincenty_length(&self) -> WasmResult<FloatArray> {
-                use geoarrow::algorithm::geo::vincenty_length;
+                use geoarrow::alg::geo::vincenty_length;
                 Ok(FloatArray(vincenty_length(&self.into())?))
             }
         }
