+++ src/alg/broadcasting/linestring.rs
