+++ src/alg/broadcasting/multipoint.rs
