+++ src/alg/broadcasting/point.rs
