+++ src/alg/broadcasting/primitive.rs
