+++ src/alg/broadcasting/vec.rs
