+++ src/alg/geo/affine/translate.rs
@@ -1,5 +1,5 @@
-use crate::algorithm::broadcasting::{BroadcastablePrimitive, BroadcastableVec};
-use crate::algorithm::geo::affine::affine_transform;
+use crate::alg::broadcasting::{BroadcastablePrimitive, BroadcastableVec};
+use crate::alg::geo::affine::affine_transform;
 use crate::array::GeometryArray;
 use crate::error::Result;
 use geo::AffineTransform;
