+++ src/alg/geo/area.rs
