+++ src/alg/geo/chamberlain_duquette_area.rs
