+++ src/alg/geo/convex_hull.rs
