+++ src/alg/geo/distance/euclidean.rs
