+++ src/alg/geo/is_empty.rs
