+++ src/alg/geo/simplify.rs
