--- src/algorithm/geodesy/mod.rs
@@ -1,4 +1,0 @@
-mod reproject;
-
-pub use geodesy::Direction;
-pub use reproject::reproject;
